P4-mq-sub-agent/requirements.txt
paho-mqtt
matplotlib
requests
json5
